# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Malamh — Facial Consent Registry API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── User ────────────────────────────────────────────────────────────────────
...
  /consent/tokens:
    get:
      operationId: listConsentTokens
      tags: [consent]
      summary: List pending consent token requests for the current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsentToken"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /consent/decision:
    post:
      operationId: consentDecision
      tags: [consent]
      summary: Approve or deny a consent token request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentDecisionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /consent/status/{token}:
    get:
      operationId: getConsentStatus
      tags: [consent]
      summary: Get consent token status (public polling endpoint)
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentStatusResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /consent/approve/{token}:
    get:
      operationId: getConsentApprovalPage
      tags: [consent]
      summary: Get consent approval page data (face owner action)
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentApprovalData"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Activity ─────────────────────────────────────────────────────────────────
...
  /billing/subscription:
    get:
      operationId: getSubscription
      tags: [billing]
      summary: Get current user subscription
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /billing/subscribe:
    post:
      operationId: subscribe
      tags: [billing]
      summary: Subscribe to a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscribeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /billing/portal:
    post:
      operationId: billingPortal
      tags: [billing]
      summary: Access billing portal or demo downgrade
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingPortalResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Monitor ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    # ─── Auth ────────────────────────────────────────────────────────────────────
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
      required: [email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        notifyOnScan:
          type: boolean
        notifyOnConsent:
          type: boolean
        notifyOnApiCheck:
          type: boolean
        createdAt:
          type: string
          format: date-time
        subscription:
          $ref: "#/components/schemas/Subscription"
          nullable: true
      required: [id, email, notifyOnScan, notifyOnConsent, notifyOnApiCheck, createdAt]
...
    ConsentToken:
      type: object
      properties:
        id:
          type: string
        faceId:
          type: string
        requesterName:
          type: string
        purpose:
          type: string
          nullable: true
        token:
          type: string
        approved:
          type: boolean
        used:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, faceId, requesterName, token, approved, used, expiresAt, createdAt]
    ConsentDecisionBody:
      type: object
      properties:
        token:
          type: string
        decision:
          type: string
          enum: [approve, deny]
      required: [token, decision]
    ConsentStatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, denied, expired]
      required: [status]
    ConsentApprovalData:
      type: object
      properties:
        token:
          $ref: "#/components/schemas/ConsentToken"
        faceLabel:
          type: string
          nullable: true
        isOwner:
          type: boolean
      required: [token, isOwner]

    # ─── Activity ─────────────────────────────────────────────────────────────────
...
    Subscription:
      type: object
      properties:
        id:
          type: string
        plan:
          type: string
          enum: [FREE, MONITOR, MONITOR_PRO]
        status:
          type: string
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, plan, status, createdAt]
    SubscribeBody:
      type: object
      properties:
        plan:
          type: string
          enum: [MONITOR, MONITOR_PRO]
      required: [plan]
    SubscribeResponse:
      type: object
      properties:
        success:
          type: boolean
        checkoutUrl:
          type: string
          nullable: true
        plan:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
      required: [success]
    BillingPortalResponse:
      type: object
      properties:
        success:
          type: boolean
        portalUrl:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
      required: [success]

    # ─── Monitor ──────────────────────────────────────────────────────────────────
...